Update issue templates
<a href="https://ijcanoncouk.medium.com/canon-pixma-ts3122-wireless-setup-mac-windows-c96e4b4e4dc3
">IJ Start Canon TS3122 Setup</a> Wireless For Mac and Windows​​ Press the Stop button present on the control board if the ON light is blazing. Then, press the Direct catch and hold it for a couple of moments. When the association is set up, the Wi-Fi image will begin flickering.